Ep7-Completed - Multiplayer P2P Lobbies
Implemented peer to peer lobbies, the ability for users to create a lobby of there own and name it.

Had trouble using Vertical / Horizontal Layouts when used together.  One would affect the other and prevent resizing.  It's a bit of a minefield.  Probably better designing the fluid UI myself next time and just offsetting from the corners of the screen as required.